requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
python-docx
streamlit
